Migrate DropInInternalClient unit tests.
